Restructure everything.
This doesn't mean it works yet.
