+++ src/bus/bus.h
@@ -1,0 +1,7 @@
+#ifndef BUS_H
+#define BUS_H
+
+#include <stdint.h>
+
+#endif
+
