--- src/include/BusMessage.hpp
@@ -1,14 +1,0 @@
-#ifndef BUSMESSAGE_HPP
-#define BUSMESSAGE_HPP
-
-enum class BusMessage {
-    DATO,
-    DATOB,
-    DATI,
-    DATIP,
-    MSYN,
-    SSYN,
-    CLEAR
-};
-
-#endif
