+++ src/libobj/obj.c
