+++ src/libobj/obj.h
