+++ src/mem/mem.h
@@ -1,0 +1,7 @@
+#ifndef MEM_H
+#define MEM_H
+
+#define MEMBYTES 65535
+#define MEMWORDS 32767
+
+#endif
